+++ src/dexterity/localroles/tests/test_utils.py
@@ -20,11 +20,11 @@ localroles_config = {
     u"private": {
         "raptor": {
             "roles": ("Editor", "Contributor"),
-            "rel": "{'dexterity.localroles.related_parent': ('Reader',)}",
+            "rel": '{"dexterity.localroles.related_parent": ["Reader"]}',
         },
         "cavemans": {"roles": ("Reader",)},
     },
-    u"published": {"hunters": {"roles": ("Reader",)}, "dina": {"roles": ("Editor",)}},
+    u"published": {"hunters": {"roles": ("Reader",)}, "dina": {"roles": ["Editor"]}},
 }
 
 
@@ -179,7 +179,7 @@ class TestUtils(unittest.TestCase):
         )
         self.assertEqual(
             dic["private"]["raptor"]["rel"],
-            "{'dexterity.localroles.related_parent': ('Reader',)}",
+            '{"dexterity.localroles.related_parent": ["Reader"]}',
         )
         self.assertListEqual(dic["private"]["cavemans"]["roles"], ["Reader"])
         self.assertEqual(dic["private"]["cavemans"]["rel"], "")
@@ -193,7 +193,7 @@ class TestUtils(unittest.TestCase):
         )
         self.assertEqual(
             dic["private"]["raptor"]["rel"],
-            "{'dexterity.localroles.related_parent': ('Reader',)}",
+            '{"dexterity.localroles.related_parent": ["Reader"]}',
         )
         self.assertListEqual(
             dic["private"]["cavemans"]["roles"], ["Reader", "Reviewer"]
@@ -207,7 +207,7 @@ class TestUtils(unittest.TestCase):
                 "private": {
                     "cavemans": {"roles": ("Reviewer",)},
                     "dina": {"roles": ("Reader", "Reviewer")},
-                    "raptor": {"rel": "{'a_utility': ('Reader',)}"},
+                    "raptor": {"rel": "{'a_utility': ['Reader']}"},
                 },
                 "published": {
                     "hunters": {"roles": ("Reviewer",)},
@@ -218,7 +218,7 @@ class TestUtils(unittest.TestCase):
         self.assertListEqual(
             dic["private"]["raptor"]["roles"], ["Editor", "Contributor"]
         )
-        self.assertEqual(dic["private"]["raptor"]["rel"], "{'a_utility': ('Reader',)}")
+        self.assertEqual(dic["private"]["raptor"]["rel"], "{'a_utility': ['Reader']}")
         self.assertListEqual(
             dic["private"]["cavemans"]["roles"], ["Reader", "Reviewer"]
         )
@@ -234,7 +234,7 @@ class TestUtils(unittest.TestCase):
                 "private": {
                     "cavemans": {"roles": ("Reviewer",)},
                     "dina": {"roles": ("Reader", "Reviewer")},
-                    "raptor": {"rel": "{'a_utility': ('Reader',)}"},
+                    "raptor": {"rel": "{'a_utility': ['Reader']}"},
                 },
                 "published": {
                     "hunters": {"roles": ("Reviewer",)},